refactor(ui): restructure into form_app/ and chat_app/ subdirectories
Move src/ui/{api_client,config,streamlit_app}.py into src/ui/form_app/.
Rename src/ui/legacy_chat_app/ to src/ui/chat_app/ (stub __init__ only —
predictor and chat_agent are replaced in the next commit). Mirror the same
split in tests/ui/. Remove legacy_chat_app exclusions from ruff and mypy in
pyproject.toml. Add anthropic>=0.40 to project dependencies.
