+++ pyproject.toml
@@ -16,6 +16,7 @@ dependencies = [
     "joblib>=1.4",
     "httpx>=0.27",
     "streamlit>=1.31",
+    "anthropic>=0.40",
 ]
 
 [project.optional-dependencies]
@@ -30,7 +31,7 @@ dev = [
 [tool.ruff]
 line-length = 100
 target-version = "py312"
-extend-exclude = ["**/legacy_*.py", "src/ui/legacy_chat_app/**"]
+extend-exclude = ["**/legacy_*.py"]
 
 [tool.ruff.lint]
 select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
@@ -40,7 +41,7 @@ ignore = [
 
 [tool.ruff.format]
 quote-style = "double"
-exclude = ["**/legacy_*.py", "src/ui/legacy_chat_app/**"]
+exclude = ["**/legacy_*.py"]
 
 [tool.mypy]
 python_version = "3.12"
@@ -49,7 +50,7 @@ explicit_package_bases = true
 warn_unused_configs = true
 disallow_untyped_defs = false
 ignore_missing_imports = true
-exclude = ["legacy_.*\\.py$", "src/ui/legacy_chat_app/.*"]
+exclude = ["legacy_.*\\.py$"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
