+++ src/ui/form_app/config.py
@@ -1,10 +1,10 @@
-"""UI configuration loaded from environment variables."""
+"""Configuration for the form-based Streamlit front end."""
 
 from pydantic_settings import BaseSettings
 
 
 class UIConfig(BaseSettings):
-    """Configuration for the Streamlit front end.
+    """Settings for the form app.
 
     All fields are overridable via environment variables of the same name
     (case-insensitive). Set API_BASE_URL in deployment to point at the
