--- src/ui/legacy_chat_app/requirements.txt
@@ -1,7 +1,0 @@
-streamlit>=1.31.0
-pandas>=2.1.4
-joblib>=1.3.2
-numpy>=1.26.3
-xgboost>=2.0.3
-scikit-learn>=1.4.0
-anthropic>=0.40.0
